Update code for Checkbox button
